feat: Finish admin event management functionality

+++ app/src/main/java/icesi/edu/co/icesicare/model/repository/EventRepository.kt
@@ -1,13 +1,11 @@
 package icesi.edu.co.icesicare.model.repository
 
 import android.net.Uri
-import android.util.Log
 import com.google.firebase.firestore.ktx.firestore
 import com.google.firebase.ktx.Firebase
 import com.google.firebase.storage.ktx.storage
 import icesi.edu.co.icesicare.model.entity.Event
 import icesi.edu.co.icesicare.model.entity.EventsHelper
-import icesi.edu.co.icesicare.model.entity.Psychologist
 import kotlinx.coroutines.tasks.await
 import java.util.UUID
 import kotlin.collections.arrayListOf
@@ -79,18 +77,15 @@ class EventRepository {
      * and that event.imageID attribute is unchanged.
      */
     suspend fun updateEvent(event:Event, isImageUpdated:Boolean){
-        Log.e("DEV","Check 0")
-        Log.e("DEV",event.toString())
         Firebase.firestore.collection("events")
             .document(event.id).get().await() //Checks event exists
-        Log.e("DEV","Check 1")
-        if(isImageUpdated){//Deletes previous image
+
+        if(isImageUpdated && event.imageId != ""){//Deletes previous image
             Firebase.storage.reference
                 .child("events")
                 .child(event.imageId).delete().await()
+            event.imageId = ""
         }
-
-        Log.e("DEV","Check 2")
 
         if(isImageUpdated && event.imageURL != ""){//Upload new image if it has one
             val uuid = UUID.randomUUID().toString()
@@ -101,13 +96,11 @@ class EventRepository {
                 .putFile(Uri.parse(event.imageURL)).await()
             event.imageId = uuid
         }
-        Log.e("DEV","Check 3")
+
         val eventF = EventsHelper.eventToEventFirebase(event)
         Firebase.firestore.collection("events")
             .document(eventF.id).set(eventF).await()
 
-        //TODO: ONLY when update succeds then change screen, show a laoding icon meanwhile
-        Log.e("DEV","Check 4")
     }
 
     suspend fun createEvent(event:Event):String{
@@ -128,11 +121,10 @@ class EventRepository {
         val uuid = UUID.randomUUID().toString()
         event.id = uuid
 
+        val eventF = EventsHelper.eventToEventFirebase(event)
         Firebase.firestore
             .collection("events")
-            .document(uuid).set(event).await()
-
-        Log.e("DEV", "MY CHECK")
+            .document(uuid).set(eventF).await()
 
         return event.id
     }
